backend/requirements.txt
scikit-learn==1.7.2
numpy==1.26.4
pandas==2.2.3
joblib==1.5.3
threadpoolctl==3.6.0
scipy==1.15.3
psycopg2-binary

fastapi
uvicorn[standard]
pydantic
python-dotenv
python-multipart
